Create new page structure
Added groups "Members" and "Events" with sub-pages in their respective
directories. In the root directory exists "main group pages" such as
`members.php` and `events.php` which provide an overview of the contents
in their subpages. Subpages are located in a directory with the same
name as the group, such as `/members/`